requirements.txt
pyTelegramBotAPI==4.12.0
requests>=2.28.0
fpdf2>=2.7.4
Pillow>=9.0.0
feedparser>=6.0.8
google-generativeai>=0.2.0
Flask==2.3.0
